ATR-441: renamed folder, analyzer and namespace
