+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/Dac/DacAnalyzersAggregator.cs
@@ -16,7 +16,7 @@ using Acuminator.Analyzers.StaticAnalysis.NonNullableTypeForBqlField;
 using Acuminator.Analyzers.StaticAnalysis.PXGraphUsageInDac;
 using Acuminator.Analyzers.StaticAnalysis.UnderscoresInDac;
 using Acuminator.Analyzers.StaticAnalysis.AutoNumberAttribute;
-using Acuminator.Analyzers.StaticAnalysis.PublicExtensionAnalyzer;
+using Acuminator.Analyzers.StaticAnalysis.NonPublicExtensions;
 using Acuminator.Utilities;
 using Acuminator.Utilities.Roslyn.Semantic;
 using Acuminator.Utilities.Roslyn.Semantic.Dac;
@@ -37,7 +37,7 @@ namespace Acuminator.Analyzers.StaticAnalysis.Dac
 			new DacNonAbstractFieldTypeAnalyzer(),
 			new ConstructorInDacAnalyzer(),
 			new UnderscoresInDacAnalyzer(),
-			new PublicGraphAndDacExtensionAnalyzer(),
+			new NonPublicGraphAndDacExtensionAnalyzer(),
 			new ForbiddenFieldsInDacAnalyzer(),
 			new DacUiAttributesAnalyzer(),
 			new InheritanceFromPXCacheExtensionAnalyzer(),
