+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/PXGraph/PXGraphAnalyzer.cs
@@ -16,7 +16,7 @@ using Acuminator.Analyzers.StaticAnalysis.ThrowingExceptions;
 using Acuminator.Analyzers.StaticAnalysis.TypoInViewDelegateName;
 using Acuminator.Analyzers.StaticAnalysis.UiPresentationLogic;
 using Acuminator.Analyzers.StaticAnalysis.ViewDeclarationOrder;
-using Acuminator.Analyzers.StaticAnalysis.PublicExtensionAnalyzer;
+using Acuminator.Analyzers.StaticAnalysis.NonPublicExtensions;
 using Acuminator.Utilities;
 using Acuminator.Utilities.Roslyn.Semantic;
 using Acuminator.Utilities.Roslyn.Semantic.PXGraph;
@@ -49,7 +49,7 @@ namespace Acuminator.Analyzers.StaticAnalysis.PXGraph
 			new NoPrimaryViewForPrimaryDacAnalyzer(),
 			new ActionHandlerAttributesAnalyzer(),
             new ActionHandlerReturnTypeAnalyzer(),
-			new PublicGraphAndDacExtensionAnalyzer(),
+			new NonPublicGraphAndDacExtensionAnalyzer(),
             new InvalidPXActionSignatureAnalyzer(),
 			new TypoInViewDelegateNameAnalyzer())
         {
